requirements.txt
# ============================================================
# macromalt 자동화 시스템 - 의존성 라이브러리
# 설치 명령어: pip install -r requirements.txt
# ============================================================

# [HTTP 요청] 웹 페이지 및 API 호출
requests==2.32.3

# [HTML 파싱] 웹 크롤링용 HTML 구조 분석
beautifulsoup4==4.12.3
lxml==5.3.0

# [RSS 파싱] RSS/Atom 피드 수집
feedparser==6.0.11

# [환경 변수] .env 파일 로드
python-dotenv==1.0.1

# [OpenAI] GPT API 연동 (기사 요약/생성)
openai>=2.0.0

# [워드프레스] REST API 연동 (자동 포스팅)
python-wordpress-xmlrpc==2.3

# [스케줄링] 정기 자동 실행
schedule==1.2.2

# [날짜/시간] 날짜 파싱 및 포맷 처리
python-dateutil==2.9.0

# [데이터 처리] JSON, CSV 등 데이터 핸들링
pandas==2.2.3

# [재시도 로직] 네트워크 오류 시 자동 재시도
tenacity==9.0.0

# [HTML→텍스트] HTML 태그 제거 후 순수 텍스트 추출
html2text==2024.2.26
